[MOV] im_livechat: rename WebsiteLivechatMessage to PublicLivechatMessage
Task-2894499
